engine: Prune based on endgame score.
If a mate is already known up to N ply from root, then there is no point in
searching at ply N+1.  Returned mate score is always number plies from
root to mate.  However, hash table still stores mate from current
position, but the score is adjusted when the position is retrieved
from cache.

The improvement is not that big because the testing framework
ends the games in which both engines agree that one side has a large
advantages. However, without this patch there is often a search
explosion when the engine finds a mate.

25208 @ 40/15+0.05
10396 - 10267 - 4545
ELO 1.78±3.88