engine: Implement late move reduction.
Late move reduction (LMR) reduces quiet moves
that are unlikely to raise α because they are very far in
the list.

LMR requires a good move ordering. So maybe I should experiment
again with IID.

LMR has more benefits when we search at higher depths, but
on ARM with 40/45 reached depth is about 7-8 which is not
enough. For further tests I'll need to increase the time to 40/60+0.05.

The conditions are not yet tuned. It's just a basic LMR.

5200 @ 40/45+0.05
2123 - 1927 - 1150
ELO 13.10±8.34