zobrist: Make zobrist key equal to polyglot key.
Use polyglot numbers to compute the zobrist key.
Enpassant is computed slightly different: it's set only
if there is an enemy pawn able to take pawn in the next move.
This change will allow zurichess to use a polyglot opening
book in the future.

No functional changes expected, but hashes changed.

910 @ 40/45+0.05
365 - 348 - 197
ELO 6.49±20.01