basic: Speedup Bitboard.Has.
On AMD64, the Go shifts operators produce a series
of instructions CMPQ/SBBQ/ANDQ to make sure that the shift
has defined behavior when the counter is larger than
the bit-width. In C a large counter is undefined behavior
so the extra instructions are not generated.

To fix this we add a &63 so that the compiler knows that
the counter is small enough and shouldn't not generate
the extra machine code. On ARM the &63 operation is redundant.

Speeds up ECM at depth 8 from 40.6s to 40.1s.