uci: Fix isready.
The UCI protocol says

-----
This command is also required once before the engine is asked to do any search
to wait for the engine to finish initializing.
This command must always be answered with "readyok" and can be sent also when the engine is calculating
in which case the engine should also immediately answer with "readyok" without stopping the search.
-----

Since initialization is very cheap and there are no table bases,
for zurichess isready should always return isready immediately.

Fixes https://bitbucket.org/zurichess/zurichess/issues/16/no-readyok-after-isready-during-search